+++ packages/kit/types.d.ts
@@ -1,5 +1,5 @@
 import './types.ambient-modules';
-import { Headers, LoadInput, LoadOutput, Logger } from './types.internal';
+import { Headers, ErrorLoadInput, LoadInput, LoadOutput, Logger } from './types.internal';
 import { UserConfig as ViteConfig } from 'vite';
 
 export type Config = {
@@ -96,16 +96,21 @@ export type RequestHandler<Context = any, Body = unknown> = (
 
 export type Load = (input: LoadInput) => LoadOutput | Promise<LoadOutput>;
 
+export type ErrorLoad = (input: ErrorLoadInput) => LoadOutput | Promise<LoadOutput>;
+
 export type GetContext<Context = any> = (incoming: Incoming) => Context;
 
 export type GetSession<Context = any, Session = any> = {
 	({ context }: { context: Context }): Session | Promise<Session>;
 };
 
-export type Handle<Context = any> = (
-	request: Request<Context>,
-	render: (request: Request<Context>) => Response | Promise<Response>
-) => Response | Promise<Response>;
+export type Handle<Context = any> = ({
+	request,
+	render
+}: {
+	request: Request<Context>;
+	render: (request: Request<Context>) => Response | Promise<Response>;
+}) => Response | Promise<Response>;
 
 export type Page = {
 	host: string;